erlang-eunit.el: Require cl, as it is used
erlang-eunit.el (part of erlang-mode) uses "flet", which is only available
if emacs package cl was loaded. The docs of the CL package suggest that
'...packages installed in Emacs must not load "CL" at run time.' We
honour this by using eval-when-compile.

This fixes warnings when byte-compiling erlang-eunit.el with:
 emacs -batch -f batch-byte-compile erlang-eunit.el

The warning fixed is:
	In erlang-eunit-compile-and-run-tests:
	erlang-eunit.el:146:4:Warning: `(save-some-buffers (&optional any) nil)'
	is a malformed function

This bug was reported initially against Ubuntu packages, see
https://bugs.launchpad.net/bugs/541893 for reference

Signed-off-by: Ralf Doering <rdoering@netsys-it.de>