inferior-erlang: specify command to run, uniquify buffer names
Add a new feature to the command inferior-erlang: when called with a
prefix argument, prompt for the command to execute to start the Erlang
shell.  When possible (i.e., in Emacs 23 and later), allow tab
completion as in a normal shell.

Also create the shell buffer in such a way that, if
uniquify-buffer-name-style is not nil and several Erlang shell buffers
are started in different directories, the names of those buffers
reflect their working directories instead of being *erlang*,
*erlang*<2>, etc.