Set erlang-mode-map when loading erlang.el
Create and assign the erlang-mode keymap to erlang-mode-map once
erlang.el is loaded, not the first time it is used.  This way of doing
it follows the Emacs Lisp convention described at
http://www.gnu.org/software/emacs/manual/html_node/elisp/Tips-for-Defining.html
and also makes it easier for users to customize the keymap.

Remove the now unneeded functions erlang-keymap-init and
erlang-mode-commands.

Also move the definition of inferior-erlang-use-cmm, so it is
available when defining erlang-mode-map.