Declare indentation options as "safe" in erlang-mode for Emacs
Emacs has a facility for setting options on a per-file basis based on
comments in the source file.  By default, all options are considered
"unsafe", and the user is queried before the variable is set.  This
patch declares the variables erlang-indent-level, erlang-indent-guard
and erlang-argument-indent to be safe, if the value specified in the
source file is valid.

Such declarations usually look like this:

%% -*- erlang-indent-level: 2 -*-

and appear on the first line of the file.