Inhibit electric newline after "->" when inside a type spec
The Erlang mode for Emacs inserts a newline after every "->", which
saves you one keystroke when writing a function, but that is
inappropriate when writing a type spec, as you'd normally keep the
spec on one line.  This change inhibits the automatic insertion when
the current line starts with "-spec" or "-type".