Remove Flymake dependency in erlang-pkg.el
The dependency was incorrect, as the package is called flymake, not
flymake-mode.  It also used to suggest a newer version of flymake than
the one that comes with Emacs, even though the Erlang mode works well
with the older version.

Removing the dependency instead of depending on version 0.3 (the one
shipped with Emacs), since the downloadable version of package.el for
Emacs 23 and earlier doesn't recognise Emacs built-in libraries as
packages.  (Thanks to Steve Purcell for pointing this out.)